+++ src/interval.rs
@@ -1,29 +1,29 @@
 pub struct Interval {
-    pub min: f64,
-    pub max: f64,
+    pub min: f32,
+    pub max: f32,
 }
 
 impl Interval {
     pub fn empty() -> Self {
         Self {
-            min: f64::MAX,
-            max: f64::MIN,
+            min: f32::MAX,
+            max: f32::MIN,
         }
     }
-    pub fn new(min: f64, max: f64) -> Self {
+    pub fn new(min: f32, max: f32) -> Self {
         Self { min, max }
     }
-    pub fn size(&self) -> f64 {
+    pub fn size(&self) -> f32 {
         self.max - self.min
     }
-    pub fn contains(&self, x: f64) -> bool {
+    pub fn contains(&self, x: f32) -> bool {
         self.min <= x && x <= self.max
     }
-    pub fn surrounds(&self, x: f64) -> bool {
+    pub fn surrounds(&self, x: f32) -> bool {
         self.min < x && x < self.max
     }
 
-    pub fn clamp(&self, x: f64) -> f64 {
+    pub fn clamp(&self, x: f32) -> f32 {
         if x < self.min {
             self.min
         } else if x > self.max {
@@ -35,11 +35,11 @@ impl Interval {
 }
 
 pub static EMPTY_INTERVAL: Interval = Interval {
-    min: f64::MAX,
-    max: f64::MIN,
+    min: f32::MAX,
+    max: f32::MIN,
 };
 
 pub static UNIVERSE_INTERVAL: Interval = Interval {
-    min: f64::MIN,
-    max: f64::MAX,
+    min: f32::MIN,
+    max: f32::MAX,
 };
